Fix Invalid Cast Exception on iOS
I upgraded from 1.0.3 to 1.1.0 and suddenly was getting a `System.InvalidCastException` being thrown from this effect.

Here is the stack trace:
```
 at Xamarin.CommunityToolkit.iOS.Effects.RemoveBorderEffect.get_TextField () [0x00006] in <ed56dc24dfeb4b7cbd0f7272b792096c>:0 
  at Xamarin.CommunityToolkit.iOS.Effects.RemoveBorderEffect.OnAttached () [0x00000] in <ed56dc24dfeb4b7cbd0f7272b792096c>:0 
  at Xamarin.Forms.Effect.SendAttached () [0x00009] in D:\a\1\s\Xamarin.Forms.Core\Effect.cs:54 
  at Xamarin.Forms.RoutingEffect.SendAttached () [0x00000] in D:\a\1\s\Xamarin.Forms.Core\RoutingEffect.cs:29 
```

Although the git history doesn't show my in here, I added this effect back to the toolkit in #492 and I originally cast the `Control` using `as` instead of a hard cast. This was to guard from throwing this exception incase the effect was applied to the wrong element ([link](https://github.com/xamarin/XamarinCommunityToolkit/pull/492/files)). In my case I use this effect on `Entry` & `Editor` elements and I now need to rollback my upgrades because many of my `Editor` elements use this effect. On Android this effect will affect `Editor` elements so I believe the code should be reverted to its original state when added 😄